Fix issue #1: Added college background
