+++ templates/success.html
@@ -1,0 +1,53 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Success</title>
+    <style>
+        body {
+            font-family: 'Arial', sans-serif;
+            background: url('/static/background.png') no-repeat center center fixed; /* Local Image URL */
+            background-size: cover;
+            display: flex;
+            justify-content: center;
+            align-items: center;
+            height: 100vh;
+        }
+
+        .confirmation-container {
+            background-color: rgba(255, 255, 255, 0.85);
+            padding: 20px;
+            border-radius: 10px;
+            box-shadow: 0 4px 10px rgba(0, 0, 0, 0.1);
+            width: 400px;
+            text-align: center;
+        }
+
+        .confirmation-container h2 {
+            color: #28a745;
+        }
+
+        button {
+            padding: 10px 20px;
+            background-color: #007bff; /* High contrast button color */
+            color: white;
+            border: none;
+            border-radius: 5px;
+            cursor: pointer;
+            font-size: 16px;
+        }
+
+        button:hover {
+            background-color: #0056b3; /* Darker button hover */
+        }
+    </style>
+</head>
+<body>
+    <div class="confirmation-container">
+        <h2>Form Submitted Successfully!</h2>
+        <p>Your data has been saved.</p>
+        <button onclick="window.location.href='/'">Back to Form</button>
+    </div>
+</body>
+</html>
